+++ tests/FrequenciesTest.php
@@ -10,3 +10,35 @@ it('can calculate frequencies', function () {
     expect($a[92])->toEqual(2);
     expect($a)->toHaveCount(11);
 });
+
+it('can calculate relative frequencies', function () {
+    $s = Statistics::make(
+        [3,4,3,1]
+    );
+    $a = $s->relativeFrequencies();
+    expect($a[3])->toEqual(50);
+    expect($a)->toHaveCount(3);
+    expect($s->originalArray())->toHaveCount(4);
+});
+
+it('can calculate cumulative frequencies', function () {
+    $s = Statistics::make(
+        [3,4,3,1]
+    );
+    $a = $s->getCumulativeFrequences();
+
+    expect($a[3])->toEqual(3);
+    expect($a)->toHaveCount(3);
+    expect($s->originalArray())->toHaveCount(4);
+});
+
+it('can calculate cumulative relative frequencies', function () {
+    $s = Statistics::make(
+        [3,4,3,1]
+    );
+    $a = $s->cumulativeRelativeFrequencies();
+
+    expect($a[3])->toEqual(75);
+    expect($a)->toHaveCount(3);
+    expect($s->originalArray())->toHaveCount(4);
+});