+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Initial release with:
 - frequencies(): a frequency is the number of times a value of the data occurs;
 - relativeFrequencies(): a relative frequency is the ratio (fraction or proportion) of the number of times a value of the data occurs in the set of all outcomes to the total number of outcomes;
 - getCumulativeFrequences(): is the accumulation of the previous relative frequencies.;
-- getCumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.
+- cumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.
 
 ## 0.1.0 - 2022-01-08
 
@@ -30,4 +30,4 @@ Initial release with:
 - frequencies(): a frequency is the number of times a value of the data occurs;
 - relativeFrequencies(): a relative frequency is the ratio (fraction or proportion) of the number of times a value of the data occurs in the set of all outcomes to the total number of outcomes;
 - getCumulativeFrequences(): is the accumulation of the previous relative frequencies.;
-- getCumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.
+- cumulativeRelativeFrequencies(): is the accumulation of the previous relative ratio.